Changes README to cover project scope and context
Adjusts README to include links to CDIPS and AFSIS
and a statement of the project goal.

Moves information about dependencies into a
requirements.txt file that is compatible with
both Anaconda and virtualenv/pip.
See their docs for how to use it.

The README could still be substantially improved,
for example to better explain the value proposition
for spectral methods.